Moving programming interface of downloader for map_repository unit to storage unit.
