--- storage/map_repository.cpp
@@ -1,6 +1,0 @@
-#include "storage/map_repository.hpp"
-
-namespace storage
-{
-MapRepository::MapRepository() {}
-} //  namespace storage
